+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Disabled the post-authentication `EHLO` capability refresh by default.
+
 ## [0.2.0] - 2026-07-15
 
 ### Added